Small refactor of metadata bits
This patch performs a small refactor of the metadata port creation and
the code which waits for metadata service to be provisioned in the
network when a port transitions to 'up'.

Also, when metadata is disabled, an error was logged trying to retrieve
fetch the metadata port. Instead of checking if metadata is enabled or
not in many points of the code, that function will return None if
service is disabled.

Change-Id: Icd7db235cff8db07f2f4e460859c71b0b4e937dc
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>