Fix the stable/pike branch
This is a squash of the following two changes:
1. I78b054398e3c2c74f84e373405d4a8fbd428d623
2. Icd7db235cff8db07f2f4e460859c71b0b4e937dc
3. I62077d750d1cfa7dbcde3e9d86228bad3e80f894

Right now, stable/pike CI is broken due to 1 missing but this change
depends on 3 which cannot be merged until we have 1. This
interdependency is solved by squashing all these changes.

* Commit 1:
Create the metadata port with the same project_id as its network

Right now, metadata_port was created with an empty project_id.
Due to this patch in Neutron [0], if the port has an empty project_id,
the default SG is tried to be recreated. This is not desirable and
when rally jobs are run it causes a DB failure because of so many
retries.

This patches fixes it by creating the metadata port within the same
project as its network instead of leaving it empty.

[0] https://review.openstack.org/#/c/507926/

Closes-Bug: #1721480
(cherry picked from commit 464430204e487ee9d5bd1a0047d9b89f03f32802)

* Commit 2:
Small refactor of metadata bits

This patch performs a small refactor of the metadata port creation and
the code which waits for metadata service to be provisioned in the
network when a port transitions to 'up'.

Also, when metadata is disabled, an error was logged trying to retrieve
fetch the metadata port. Instead of checking if metadata is enabled or
not in many points of the code, that function will return None if
service is disabled.

* Commit 3:
Move networking-ovn legacy jobs to our tree

This patch is moving the legacy CI jobs for networking-ovn into our tree
instead of having them as part of the openstack-infra/openstack-zuul-jobs.

Zuul v3 allows projects to do that in order to give them more control
over their jobs.

By having these jobs in the networking-ovn tree also facilitates the
eventual migration from legacy jobs to the new ansible roles.

Change-Id: I62077d750d1cfa7dbcde3e9d86228bad3e80f894
(cherry picked from commit f1eada645ddabdf140d793e63d3cf9c7f9693df5)