Support ACL name and severity columns
New fields 'name' and 'severity' added in OVN ACL table for logging
purpose. In this change, we are updating Networking-ovn with
these fields(only if 'ACL' table supports them to maintain
Networking-ovn compatibility with other release branches of OVS,
which won't have these columns)

Note: Earlier gate functional tests were broken when these fields were
introduced in ACL table. As a temporary fix we pushed commit [1], which
forced functional tests to run on ovs branch 2.7. Now we are reverting
that as we are proposing proper fix with this patch.

[1] e951289b3664851e64952b31ff70043a01dd2cb0
Closes-bug: #1708594

Change-Id: I580a4c80d875449d6c2571f439ffe3e8f1d46861