Fix gate mtu tests
This new commit [1] in neutron checks mtu size during segment
allocation and throws exception if MTU > 1450 for geneve network
type. Because of this new change, tests
'test_clear_segment_host_mapping' and
'test_update_segment_host_mapping' are failing in gate.

To fix this, we are updating mtu to 1450 in the test, before it
creates geneve segments for the network.

[1] ceb86366d25a928d7a3773075a7333801ee3dd29

Closes-bug: #1710496
Change-Id: I117cc2352848718b37f95bdb6aecbfb6ba1ccc67