Add mac from allowed_address_pairs to ovn lport addresses
Currently allowed_address_pairs works only if the mac in the allowed
address pair is the same as the mac address of the port (ip can be
different). Otherwise, packets destined to the new mac address
cannot be delivered to the port, because there is no logical flow
in L2 lookup stage in OVN for the new mac. This patch fixes it by
adding the new mac to "addresses" of the OVN lport.

Change-Id: I69e85f2a1acb9541e624c8a153610f9759b8d238
Closes-bug: #1711887