Enhance devstackgaterc to support neutron-api tests
This patch is modifying the devstackgaterc file to add support for
neutron-api tests.

Before this patch, devstackgaterc only supported 1 positional argument
for choosing the ovs branch (latest-release or master). Now, it possible
to pass as many options as needed and it will be much easier to extend.

This patch will be used at: 6791b94429f4a2f494e3f45426e074e8608ce917

Change-Id: Idee833a90c2c2dacea8c92ff58f67f6c3822bb9e