neutron-ovn-db-sync-util: sync metadata ports
When migrating from ML2/OVS or OVN without metadata to an OVN
environment with metadata support, the existing networks must have
a metadata port in it with an IP address on each of its subnets.

This patch ensures that all networks have such port. Also, in case
that a network had more than one DHCP port, it will delete all but
one, which will be used for metadata (and eventually for DHCP) in
OVN.

Change-Id: I4c9fa6c08a489877c85d7bb725009ab5cd5f8120
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>