Don't create metadata port if it already exists
When creating a network in OVN a metadata port was created
unconditionally. In some cases like when doing a DB sync, it may happen
that a network had been deleted from OVN but both the network and
its metadata port still exists in Neutron. In this case, the sync
process will recreate the network in OVN and a second metadata port
would be created again.

This patch fixes it by looking if the metadata port already exists.

Change-Id: I8ae32e131d34f132b55b956119f78e6ab9fe9972
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>