Create the metadata port with the same project_id as its network
Right now, metadata_port was created with an empty project_id.
Due to this patch in Neutron [0], if the port has an empty project_id,
the default SG is tried to be recreated. This is not desirable and
when rally jobs are run it causes a DB failure because of so many
retries.

This patches fixes it by creating the metadata port within the same
project as its network instead of leaving it empty.

[0] https://review.openstack.org/#/c/507926/

Change-Id: I78b054398e3c2c74f84e373405d4a8fbd428d623
Closes-Bug: #1721480
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>