Remove "fake_api" from functional tests
The mocking of the API is not necessary and prevents the use of
built-in functions like api.db_find() in code that is tested. This
change is necessary to move over to using ovsdbapp, as many of the
direct accesses to the backend-specific idlutils will be replaced
by using common db access methods.

Change-Id: I0e195e62f943d9fdc4dafadc37f4870a392f6eb2