Simplify the L3 {create, update}_router() methods
This patch is moving some of the logic from the {create,
update}_router() methods of the L3 driver into the OVNClient. In
summary, two things is being done in this patch:

* Remove the "networks" parameter from the {create, update}_router()
  method in OVNClient. Now the list of network is fetched within these
  methods.

* Remove the "delta" parameter from the update_router() method in
  OVNClient. We are already passing the new and old version of the
  router to that method, we know the differences we don't need to have
  an extra parameter for it.

This patch is related to the work to that is being done at bug #1605089
which uses the OVNClient to recovery the resources that are out of sync
and by having OVNClient to handle the bulky work does simplify the
synchronization work.

Related-Bug: #1605089
Change-Id: I8f0afb841e042f75e5062cca81717a3ecad17015