Log a warning when trying to wait for metadata on a non existent port
As noticed in bug #1728088, we're receiving an 'up' event from a port
that does not exist. Even though we may want to prevent this from
happening, the current result is that we're throwing a confusing
exception because the chassis and datapath for this logical port can't
be retrieved.

This patch detects the above condition and logs a warning instead.

Related-Bug: #1728088

Change-Id: If15f457db6a99ba232e23349048fcbe2b1a4f36d
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>