Tempest test_port_security_macspoofing_port was skipped for wrong reason
According to the comment, this test was skipped because it expected
the dhcp interface to be present which is not true. The reason why this
test has to be skipped is because the connection is checked through a
ping command from the spoofed port and replies won't reach the
destination because of the OVN ARP responder: ICMP replies will be sent
to the original MAC address and not to the spoofed one.

Let's continue to skip it until [0] gets fixed.

[0] https://bugs.launchpad.net/tempest/+bug/1728886

Change-Id: I3a8d2c04f34eab241350e88f3e0b6f57d3dd37ca