Check for correctness when updating networks
This patch is saving the revision number from the Neutron network object
into OVNDB and using it for checking for correctness when updating the
OVN resource.

The "original_network" parameter has been removed from the
update_network() methods. Since the stragey here is to drop patches with
older revision numbers we won't be creating a delta between two neutron
versions anymore. Instead, we will always make sure that OVN is sync'd
with what is latest in Neutron.

Partial-Bug: #1605089
Change-Id: Ie817a559bd708f6601d7004187f0f088e8b09b47