Optimize inconsistency detection (Part 1)
Create the database model and migration scripts for saving the revision
number of the OVN objects in the Neutron database.

Update the methods related to the network resource to use the new table.

Next patches in the series will use this new table to detect when the
resource is out of sync and fix it.

Related-Bug: #1605089
Change-Id: Id14e9e8b88797acec11b72117c6985c49ea713c9