Optimize inconsistency detection (Part 2)
This patch is introducing the maintenance mechanism for the database
consistency work.

This new mechanism is based on the futurist library [0].

The ``check_for_incosistencies`` task will compare the
"ovn_revision_numbers" table with the Neutron's "standardattributes" table
and check if there's any difference between the revision numbers. If
some inconsistency is found, the resource will be sync'd with the
latest version in neutron. A OVSDB lock is used for this task as well
to guarantee that only one check is running in the entirely cluster.

The maintenance mechanism runs on its own OVN worker [1] so the OVSDB
lock doesn't conflict with the OVSDB monitor one.

Currently in this patch, only the ``network`` resource is being
addressed to keep it small and easy to review. New resources will be
added on subsequently patches.

[0] https://github.com/openstack/futurist/
[1]
https://docs.openstack.org/networking-ovn/latest/contributor/design/ovn_worker.html

Change-Id: I61757e70460b83b93429ddf4f7792af7cc2d7ab5