Fix race condition on ovn_client on startup
On startup, when we initialize the IDL we start receiving notifications
right away and processing them. There's a possibility that even though
the IDLs are initialized, post_fork_initialize has not yet created the
ovn_client object resulting in an exception (and also possible missing
DB updates) if attempted to be used.

This patch fixes this by converting ovn_client object in mech driver
into a property which initializes it if doesn't exist yet.

Change-Id: I3965ae61fb5ba7dda07bb797a09383a6c35f589e
Closes-Bug: #1731288
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>