Add native DNS support
OVN supports native DNS. This patch makes use of this feature for
internal DNS resolution. With this, VMs within the same network can use
the 'dns_name' associated with the other ports. When the VMs send the
DNS request, it is resolved by ovn-controller if the dns is configured
in the 'DNS' table in the OVN Northbound db.

For each network, a DNS row is created and the 'records' column contain
the dns names and the ip addresses to resolve. The dns record is added
only of the 'dns_name', 'dns_assignment' and 'device_id' of the port is
set.

In order to use this feature, "dns" extension driver should be loaded
and 'dns_domain' in neutron.conf should be defined other than 'openstack.local'.

Change-Id: I59bdc0bc2665118fe8025a12a1c1961e04444abc
Closes-bug: #1688172