Collect functional test logs and add logstash index
Currently we don't have functional test logs in the gate which
makes hard to debug failures.

This patch aims to collect logs for those tests in the same
way that Neutron does. It also generates an index file containing
oslo.log messages. This file will then be used to populate logstash
index.

Change-Id: I485425415f5faf01a4ab45ac4ec9aeb78dfeda58
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>