Cleanup _add_router_ext_gw()
This patch is cleaning up the _add_router_ext_gw() method. After the
commit 8cc1330884f94c29d78b6bb42283ebf2f36e4ee5, all creation and update
of those resources will be done in a single transaction so, having a
try..except block around the helper functions to be able to call
_delete_router_ext_gw() doesn't work, because the commit() will be
invoked later. Plus, if something else fails on those helper methods the
whole transaction will be aborted and nothing will be created in the
first place.

Change-Id: Ic27fe8059dd851e2189642e08d639e50fdf3d787