Refactor subnet dhcp options methods
This patch is refactoring some the subnet dhcp options methods to make
it simpler to implement the database sync work [0].

The get_subnet_dhcp_options() method has extended to also be able to
return the dhcp options referent to the ports with a new
``with_ports`` parameter.

The return value of the get_subnet_dhcp_options() also changed to a
dictionary containing two keys: ``subnet`` and ``ports`` where each
value is the DHCP_Options referent to those resources.

On the update_subnets() method, start the transction on the higher level
method and have the helper methods to just add the OVSDB commands to it.
This sill later simplify the place where the revision_number comparison
will happen (at the higher level method).

Two methods are no longer needed and have been removed:

* get_subnet_and_ports_dhcp_options
* compose_dhcp_options_commands

[0] https://review.openstack.org/#/c/490834/

Change-Id: Ie930fc448d65545ed1a39d15a55c6ead563095e5