Remove redundant SetLSwitchPortCommand in trunk driver
self.plugin_driver._plugin.update_port is invoked by the OVN trunk driver,
which actually is update_port in the ml2 plugin.
update_port invokes mechanismDriver callback and then OVN updates lsp.
Therefore, the separate set_lswitch_port in OVN trunk driver is redundant,
it should be deleted.
Because of this change, unit test cases need to be adapted.
And patch[1] is squashed into this patch to avoid conflict.
[1]https://review.openstack.org/#/c/526407/

Change-Id: I202bc72e6978bc532a00a5a7c9161de3f85a9324