Refactor Floating IPs related methods
This patch is bringing logic for handling floating ips out of the L3
driver and into OVNClient.

The code now uses the ``external_ids`` column in the NAT table from
OVNDB to bind the FIP from Neutron DB with it (before we used a
combination of external + logical ips to identify those entries).

The values added in the ``external_ids`` are:

* neutron:fip_id = The ID of the FIP in neutron

* neutron:fip_port_id: This is used in the update_floatingips() method
  to compare and see whether the port_id of the FIP has changed in the
  update or not.

* neutron:router_name: The name of the router which the FIP is associated
  with.

The ovn_db_sync.py script was updated to make use the updated methods
in OVNClient to create and delete floating IPs, that way we avoid the
code duplication that existed between this script and the L3 driver.

Partial-Bug: #1605089
Change-Id: I21e9728c93392de4c579ea00c34059ad6812f682