Initialize privsep in networking-ovn-metadata-agent
We were not initializing privsep in the metadata agent so we weren't
using any root helper to call it. In devstack this is not a problem
since 'stack' user can execute anything but in TripleO deployments
it won't work since user 'neutron' is only allowed to execute commands
through rootwrap/rootwrap-daemon. This patch initializes privsep which
whatever root_helper is configured.

Change-Id: I8b1552f5adfef4f3fb114445b463a4c77b25ac9c
Closes-Bug: #1737587