Refactor security groups
This patch does two things:

1. Delete the security group at the AFTER_DELETE notification level

Prior to this patch we were using the BEFORE_DELETE notification level
which can be problematic because if the deletion fails in Neutron we
would have deleted it from OVN already.

2. Remove security groups updates

The only thing that the security group updates did was to update the
external_ids of the Address_Set with its current name.

That name wasn't used for anything in networking-ovn other than knowing
whether the security group was created by networking-ovn or not.

So, instead of having the name saved in the external_ids this patch
changes it to the security group ID which is the canonical identifier of
the resource and won't change.

Partial-Bug: #1605089
Change-Id: I5a8b64639d38e11050dae1008b14d48de14ecf94