Fix RDO OVN scenario job by checking DB schema
This commit [0] added the external_ids column to the NAT table
in OVN NB database. As it's not yet included in any OVS release,
this patch checks for the column existence before writing data
into it. Since the existing code is backwards compatible, it
will still work without that column.

Once we're sure that the minimum required OVS version contains this
column, we can revert this patch.

[0] https://github.com/openvswitch/ovs/commit/c4c1e1a5af31c65c7b78e37973d5113f64339701

Change-Id: I907ce96d2e7b99ff01646a929ac5f60a9150e52f
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>