Make ovn_client create_network idempotent
While creating a network, the call to bump the revision
in the neutron database could fail. That means that a later
healing attempt will try to recreate the network on the
ovn database.

This change makes the ovn_client.create_network idempotent
on the logical switch add call to ovsdbapp. The physnet
provnet port and metadata port creation are already designed
as idempotent.

I found no obvious place to test it, but I believe we should
do it when we migrate to ovsdbapp.

Change-Id: I20e83a0e51c8d1da3b160b2356b7232873a7a8af