Check for router port correctness
This past checks correctness when adding, removing router
interfaces, updating them, or managing the router gateway port.

For router ports, every resource (port) on the neutron side,
becomes two on the ovn side. This is why we introduce a new
type of resource to be tracked ROUTER_PORTS, and we also track
them as standard L2 PORTS.

The ROUTER_PORT resources handle the creation and update of
Logical_Router_Port and NAT resources (always going on the
same transactions).

Change-Id: I96a7e812427ebbc8e868a0e3ab78754084cdaecf