Use metadata IP as DHCP service IP for subnet without gateway IP
Currently, OVN does not support DHCP for subnet without gateway IP.
This is an important feature for VMs having multi ports on multi
subnets, because two conflicting default routes should be avoided.
In the patch[1] in OVN, router option was changed from mandatory to
optional, then we can add DHCP without router option to OVN. But,
without gateway IP, we still need an IP address as DHCP service IP,
using the same IP with metadata is a choice.
This patch does not solve all the problems, a TODO is added in the
code and bug is still opening to track it.

[1]https://github.com/openvswitch/ovs/commit/b89d25e5694bc9bb0a1fbb2a025d18a2a231c633

Change-Id: Icd93101bf8834a56f80184f0584a73a36b16b483
Partial-Bug: #1670666
Signed-off-by: Dong Jun <dongj@dtdream.com>