Update LRP when a router port is updated
Right now when a router port is updated, the corresponding
Logical Router Port is not updated. This patch checks if the
port to be updated is a LRP and updates it accordingly.

Also, I'm removing the router_id parameter in router_port_update
method since it was no longer used.

Closes-Bug: #1744087
Change-Id: I939f4c40a5e9e58c3a216353f03179578b81f64c