Maintenance task: Ordering resources by type
When fixing the creation or update of a resource, it's better to sort it
by its type starting with the root resources and leaving the leaf ones
at the end. For example, in case a network and a port failed to be
created, if the maintenance thread tries to create the port prior to
creating the network it will fail.

The same could happens to deletion, it's preferable to delete the leaf
resources prior to deleting the root ones to avoid any conflict.

This patch is adding such sorting mechanism to the maintenance.py database
module that is used by the maintenance task when listing the resources
that needs to be fixed.

Change-Id: Ibc8312bf70b221626e389ea68bfa4c69f6a2cb67
Partial-Bug: #1605089