functional tests: Register SQLAlchemy models
As the order of executed test is random, it may happen test using
database is executed first. In such case database schema is created
without ovn_revision_numbers table because SQLAlchemy model is loaded
later. It causes an attempt to remove the table during cleanup because
it's registered by the time of the cleanup.

This patch loads all the models beforehand so no race can happen. It
also fixes a case when you want to run just a single test using
database.

Change-Id: Ic6d449f4f652abd48ae1b285e0ef965c4a6329e1
Closes-bug: #1744320