Schedule gateway on chassis from ovn-cms-options
Admin sets ovn-cms-options in external_ids as

ovs-vsctl set open .
   external_ids:ovn-cms-options="enable-chassis-as-gw"

to enable a chassis as a candidate for scheduling gateway router.
Networking-ovn will parse ovn-cms-options and select this chassis
if it has proper bridge mappings.

This helps admin to exclude compute nodes to host gateway routers as
they are more likely to be restarted for maintenance operations.

We follow this order for selecting candidates
1) candidates with ovn-cms-options and proper bridge mappings
2) if no candidates from 1), then chassis with proper
   bridge mappings

Conflicts:
	doc/source/admin/refarch/refarch.rst

Closes-bug: #1743745
Change-Id: I86fbe27f0b6a9317ad82c2bcf2a0446d118de35b