Fix RDO OVN scenario job by checking DB schema (iii)
We were looking for the existence of the 'external_ids' command
in all tables. However, as floating ips and subnets are fetched
prior this check, it still fails if the OVSDB schema doesn't
include the column in any of the tables.

This patch is moving the check upwards to avoid the failure.

Closes-Bug: #1740698
Change-Id: If4c64c428d3c474f08cb0441a2c1653fb0095d37
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>