Invert order when collecting logs in tempest-post-common.yml
This patch is inverting the order of the tasks in the
tempest-post-common.yml file because, in case the job fails before
devstack ran successfully, there's a chance that the OVN databases
weren't created yet. This would make the task collecting the database
files to fail and and return before the task that collect the rest of
the logs is executed.

Change-Id: I8a5fca9203124d0d405bdf6f159b6bfecf51059c