Fix error in router port maintenance code
When a port needs to be created or updated on OVN side by the
maintenance task we need to recover the router ID from the
port, which is on the device_id field.
At some point in time during development it got changed to
device_owner, which is a logical name referring to the type
of owner: compute, router, dhcp, etc...)

Also, when getting the lrouter port from the NBDB we were
using a uuid, while we needed to convert uuid to lrp port
name.

Those mistakes was causing the maintenance code uncapable of
recovering router ports that were detected as outdated, or
not created on the OVN side.

Change-Id: Iea8060b11cc4076e6efc300dd9079ddafcc3fb5e
Closes-Bug: 1746979