Remove kuryr legacy job and move the Zuulv3 one to experimental gate
As current kuryr job is not even installing networking-ovn nor kuryr,
this patch is removing it. Even if we decide to fix it [0], its value
is still unclear since we would be depending on more projects and
we would have to run kuryr's tempest tests instead of ours.

As kuryr folks are setting up a new job to their gate using OVN as
a backend [1], let's remove this job from networking-ovn gate for now
and run their Zuulv3 job to the experimental pipeline.

Closes-bug: #1748857

[0] https://review.openstack.org/#/c/543419/
[1] https://review.openstack.org/#/c/543864/

Change-Id: Iea69d16a6ae919404987643732ee3ac6582c31a7
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>
(cherry picked from commit 71a5d899750079e11edaa84a860dfca72275143f)