Fix test_provision_path() unittest
The patch aaf11f45ecf8f832301491017a8009f1897e8d43 in Neutron changed
the way the IPDevice.exists() method work and it broke the unittest
test_provision_path() in networking-ovn.

This patch is mocking the IPDevice.exists() method now to what is
expected on that test so it doesn't rely on it's implementation anymore.

Change-Id: I179716da4d1d46012a53ec8e6b470b57de952beb