Fix UpdateLRouterPortCommand to work with OVS 2.8
The UpdateLRouterPortCommand was broken when used with OVS <= 2.8 due
to the ipv6_ra_configs column not existing in the OVNDB schema.

This patch is adding a check on that command to ignore that column if
it's not present. A TODO was left as a comment so we don't forget to
remove this check once the support for older OVS versions are dropped.

Change-Id: I1ce87173731146e3f0078df99b7ba56f44b14497
Closes-Bug: #1757438