Support per network dns domain name in internal OVN dns
If user has set network's dns_domain attribute, this patch stores an
additional record in the DNS.records column as
<port host_name>.<net_dns_domain>=<IP ADDRESSES>

This patch doesn't handle the scenario when a network's dns_domain
is updated. This is presently left as TODO and it will be supported
as a follow up patch.

Handling the DNS entries in the OVN NB DB seems to become more complicated
and I think it is better if we store the dns name of a port in
Logical_Switch_Port.options column and the dns domain name and network's domain
name in Logical_Switch.options and let ovn-northd read these values and update
the OVN SB DB 'DNS' table. I will pursue this approach before addressing the
above TODO.

Change-Id: Ic9068d4bbb4602aa9d8c5d9cce8457b39057c802
Partial-bug: #1777978
(cherry picked from commit fb1b29e74ae7bfe30894fc0a0b840da697fecaee)