Simplify the TripleO install guide
This patch is simplifying the TripleO deployment guide by:

* Removing the baremetal instructions since it's not recommended anymore
* Adding a ovn.yml file to generate the correct overcloud-deploy.sh and
  overcloud-prep-containers.sh scripts
* Clarifying some steps about deploying directly on the localhost
* Replace the workaround that was manually creating the
  network-environment.yaml file with a simpler workaround using ansible
  tags (lol, I know right?!)

Note: The tripleo/ovn.yaml could probably live in the tripleo-quickstart
repository (under config/general_config). This patch is adding it to the
networking-ovn tree for simplicity (likely we have a .sample config file
for DevStack). We can always move it later to the tripleo-quickstart
repository if we see the need for it.

Co-Authored-By: Numan Siddique <nusiddiq@redhat.com>
Change-Id: Ie38fcca8ce8334bed6acc827a136656d2c5da3d0