doc: Clarify about metadata port owner and port status
In networking-ovn, the metadata port is created under owner network:dhcp
so that it gets auto deleted when the network is removed. Also, this
port will be used to reach the DHCP when there's no gateway setup.
Compared to ML2/OVS, it can confuse operators when they still see a DHCP
port even though the network was created with the --no-dhcp flag. Plus,
as this is a distributed port (OVN localport), it will not get bound to
any chassis and it'll remain DOWN.

This patch is adding this information to the metadata design documentation.

Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>

Change-Id: Id12b57455ff30eab37a593b85a15313fafd32638