Clear binding:host_id when a subport is removed from a trunk
In networking-ovn we set the binding:host_id of a subport to its
parent's. However, when we remove the subport, that info remains there
so if that port would be reused on a trunk which is on a different host,
it will not get bound to the Chassis by ovn-controller, thus it won't
transition to 'up' and never become ACTIVE in Neutron. Tempest test
test_trunk_subport_lifecycle exercises this but the failure doesn't show
up in the gate since both servers are spawned in the same host (we don't
have multinode jobs in the gate).

This patch clears up the host_id when the subport is removed from its
trunk.

Closes-Bug: #1761194
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>
Change-Id: Iea31cb3dd418e086c43a797236c4b71d432422f7