Make tempest API/Scenario tests voting
This patch is making the tempest API and scenario tests voting in gate.
This is highly desired because a non-voting job doesn't always get the
attention it needs.

Right now only the "test_router_interface_status" is currently failing
due to a known bug which is being tracked at [0]. I've disabled it and
left a FIXME not to remove the regex once the bug is fixed.

[0] https://bugs.launchpad.net/networking-ovn/+bug/1713835

Depends-On: If5098f6e60742421022712231589afe6755878dc
Change-Id: I2226e06e5b819ff39189cc125563dc1f3d1f93af