Zuul: Avoid POST_FAILURE when OVSDB files doen't exist
If a job fails before the OVSDB files are created they are being marked
as POST_FAILURE instead of only FAILURE by the gate. This is because the
post job will try to copy the OVSDB files as part of the collection logs
task but, if the files doesn't exist the post job will also fail.

This patch is verifying if the ovnsb_db.db file exist [0] prior to trying
to copy and compact them.

This patch is also inverting the order of the post file to first check
the databases and compact it and then collect the logs. Otherwise the
databases won't show up in the logs.

[0] The ovnsb_db.db file was choosen because it's the last one of the
the db files that will be created.

Change-Id: I9f12f5113ef23c236167f3dbe53016e1351e7f13