Break schedule_unhosted_gateways() into multiple DB transactions
When we're scheduling multiple gateway ports using the Least Loaded
policy (default), we need to take into account the actual load of
all chassis after each port is scheduled. Prior to this patch, the
scheduling happened under the same DB transaction so it will result
in scheduling all ports in the chassis which was less loaded by the time
the function was called, resulting in a very likely unbalanced state of
the gateway nodes.

NOTE: I'm working on bug 1762694 which will introduce functional
tests exercising this part.

Change-Id: I4404b0958d5a835a5ec449a6ff6696d6de6c78ef
Closes-Bug: 1763005
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>