L3HA: Take priorities into account in least loaded scheduling
This patch is implementing an algorithm to take the priorities of
the gateway ports into account when choosing a chassis using the
Least Loaded policy (default). Prior to this patch, the same chassis
would always be selected leading to an unbalanced distribution
of the gateway nodes.

Now, when a new gateway is being scheduled, it will take into
account the current priorities and will select the nodes and their
priorities accordingly so that the system remains balanced even
in the event that one of the nodes go down.

Also, I'm adding a new functional test which will check that the
ports have been scheduled as expected by looking directly into the
Southbound database.

Please, note that bug 1762691 could still make routers to not
be in HA mode and we need to address this issue separately.

Closes-Bug: 1762694
Change-Id: I2df35cc31e856b9b574b77dc6b5ab5be262a12b7
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>
(cherry picked from commit bbf0260496bf2e08db3c341b88e124fa6d0c7156)