Do not use a transaction for get_logical_port_chassis_and_datapath()
The OVSDB monitor for the Northbound database is invoking
set_port_status_up(), this method will invoke another function
called _wait_for_metadata_provisioned_if_needed() which
will then invoke get_logical_port_chassis_and_datapath(). The
get_logical_port_chassis_and_datapath() starts a transaction in the OVSDB
Southbound database but, if the IDL for the Southbound database doesn't
have a valid lock [0] it will raise an exception which will prevent the
port status to be marked as UP.

Since the get_logical_port_chassis_and_datapath() is just reading from
the Southbound database we do not need to use a transaction for it and
just look for the row we want directly from the in-memory replica of the
Southbound database.

[0] https://github.com/openstack/networking-ovn/blob/df880435d3a7e98aaa8a0ba3a0a9b81286129586/networking_ovn/ovsdb/ovsdb_monitor.py#L285

Closes-bug: #1762933
Change-Id: I650c19164d619a9318b25f48dc113774f8ba67ed