Fix test_models_sync() functional test
This patch is fixing the test_models_sync() test from oslo.db by removing
the inconsistencies from the models.py module and the migration script
for the ovn_revision_numbers table.

Note that this test doesn't current run in our gate, it gets skipped
because the databases are not properly configured when setting up the
functional tests. For now, it's possible to configure mysql by hand to
test this patch by doing the following:

1. Stack DevStack as per our documentation

2. Configure Mysql:

  $ cat << EOF > /tmp/mysql.sql
  CREATE DATABASE openstack_citest;
  CREATE USER 'openstack_citest'@'localhost' IDENTIFIED BY
  'openstack_citest';
  CREATE USER 'openstack_citest' IDENTIFIED BY 'openstack_citest';
  GRANT ALL PRIVILEGES ON *.* TO 'openstack_citest'@'localhost';
  GRANT ALL PRIVILEGES ON *.* TO 'openstack_citest';
  FLUSH PRIVILEGES;
  EOF

  $ /usr/bin/mysql -u root < /tmp/mysql.sql

3. Change to the networking-ovn directory

  $ cd ~/networking-ovn

4. Run the test_models_sync() functional test:

  $ tox -edsvm-functional TestModelsMigrationsMysql.test_models_sync

Partial-Bug: #1766639
Change-Id: Ie3b600d4f3d7a3d594685848f35868c438f20976