Properly configure the environment for functional tests
This patch is configuring the functional tests environment so we can
also run the database migration tests.

The patch includes a new script at tools/configure_for_func_testing.sh
which runs as part of the gate_hook.sh for the dsvm-functional tests and
is responsible for installing and configuring the environment.

Closes-Bug: #1766639
Change-Id: I0df72a17f78c011d68ae5662e8550ce7714c1680