Lower log level for the maintenance thread
Prior to this patch the maintenance thread was logging an warning when
any inconsistencies were found between the Neutron and OVN database
without any message saying that it's going to be fixed automatically.
That caused confusion between users.

This patch is changing the message to be less "scary" and lowering it's
log level as well because the inconsistencies are going to be fixed
without the need of any manual interaction from the administrator.

DEBUG was choosen over INFO for the message because usually INFO is
choosen to indicate that something has been completed already instead
of some on-going operation. That said, we are already logging an INFO
message at the end of the synchronization to inform that it has been
completed and how long it took.

Change-Id: I520f3f1f6bd61c2ec8d8ed29260c7ebec8f0184f