Handle port's 'port_security_enabled' flag when adding ACLs
When a port's security groups are cleared and 'port_security_enabled'
flag is True, it is expected that all the traffic to/from the port is dropped.
But this is not the case. So this patch fixes this issue. 'port_security_enabled'
flag is checked before adding the port's ACLs.

So the new (and correct) behaviour with this patch is
 - No security groups and port security enabled -> Drop all the traffic
 - No security groups and port security disabled -> Allow all the traffic

Change-Id: I5265da4c21e58672db9858555625c8e2c5350316
Closes-bug: #1770723
(cherry picked from commit d97a670257fb67917e2b5da4c944b9e29a5d7855)